covid_project/kafka/consumer/requirements.txt
confluent_kafka
google-cloud-storage
pandas
